// src/main/model/entity/DataHistory.ts
import { Sequelize } from "sequelize"
import { Column, DataType, Model, Table } from "sequelize-typescript"

export interface DataHistoryAttributes {
    id: number
    modelName: string
    idModelName: string
    valueBefore: Object | null
    valueAfter: Object | null
    updatedAt: Date
    updatedBy: string
}

type DataHistoryAutoGeneratedAttributes = "id" | "updatedAt"
type DataHistoryOptionalAttributes = "valueBefore" | "valueAfter" | "updatedBy"
export type DataHistoryCreationAttributes = Omit<DataHistoryAttributes, DataHistoryAutoGeneratedAttributes | DataHistoryOptionalAttributes> & Partial<Pick<DataHistoryAttributes, DataHistoryOptionalAttributes>>

@Table({
    freezeTableName: true,
    schema: "public",
    tableName: "data_history",
    underscored: true,
    createdAt: false,
    indexes: [
        {
            name: "data_history_id_model_name",
            using: "HASH",
            fields: [
                {
                    name: "id_model_name"
                }
            ]
        },
        {
            name: "data_history_model_name",
            using: "HASH",
            fields: [
                {
                    name: "model_name"
                }
            ]
        },
        {
            name: "data_history_updated_at",
            using: "BTREE",
            fields: [
                {
                    name: "updated_at"
                }
            ]
        },
    ]
})
class DataHistory extends Model<DataHistoryAttributes, DataHistoryCreationAttributes> implements DataHistoryAttributes {
    @Column({
        allowNull: false,
        type: DataType.BIGINT,
        autoIncrement: true,
        primaryKey: true
    })
    id!: number

    @Column({
        allowNull: false,
        type: DataType.STRING
    })
    modelName!: string

    @Column({
        allowNull: false,
        type: DataType.STRING
    })
    idModelName!: string

    @Column({
        allowNull: true,
        type: DataType.JSONB,
    })
    valueBefore!: Object | null

    @Column({
        allowNull: true,
        type: DataType.JSONB,
    })
    valueAfter!: Object | null

    @Column({
        allowNull: false,
        type: DataType.DATE,
        defaultValue: Sequelize.literal("now()")
    })
    updatedAt!: Date

    @Column({
        allowNull: false,
        type: DataType.STRING,
    })
    updatedBy!: string
}

export default DataHistory